requirements.txt
# Game Image Analyzer - Requirements
# Install with: pip install -r requirements.txt

# Core image processing
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
Pillow>=10.0.0

# Data processing
pandas>=2.0.0

# Static visualizations
matplotlib>=3.7.0

# Interactive visualizations
plotly>=5.15.0

# API requests (for PokeAPI)
requests>=2.31.0

# PDF report generation
reportlab>=4.0.0

# GUI (desktop application)
customtkinter>=5.2.0
